requirements.txt
fastapi
uvicorn
python-multipart
gTTS
openai
python-dotenv
pydantic
requests
